Refactoured and Handle all Runtime Exception and validate customerServiceImpl class..

+++ Pos-Back_End/src/main/java/lk/ijse/spring/service/impl/CustomerServiceImpl.java
@@ -12,6 +12,7 @@ import lk.ijse.spring.dto.CustomerDTO;
 import lk.ijse.spring.entity.Customer;
 import lk.ijse.spring.repo.CustomerRepo;
 import lk.ijse.spring.service.CustomerService;
+import lk.ijse.spring.util.ResponseUtil;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,16 +35,25 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public void addCustomer(CustomerDTO dto) {
+        if(repo.existsById(dto.getId())){
+            throw  new RuntimeException("Customer "+dto.getId()+" Already Exists..!!!");
+        }
         repo.save(mapper.map(dto, Customer.class));
     }
 
     @Override
     public void deleteCustomer(String id) {
+        if (!repo.existsById(id)){
+            throw new RuntimeException("Customer "+id+" Not Available to Delete..!");
+        }
         repo.deleteById(id);
     }
 
     @Override
     public void updateCustomer(CustomerDTO dto) {
+        if (!repo.existsById(dto.getId())){
+            throw new RuntimeException("Customer "+dto.getId()+" Not Available to Update..!");
+        }
         repo.save(mapper.map(dto,Customer.class));
     }
